+++ examples/beachlitter/models/Model_InEqualOut.csv
@@ -1,0 +1,118 @@
+SHEETS,0,1
+SHEETS,1,2
+SHEETS,2,3
+SHEETS,3,4
+BEACHES,0,0,1,0
+BEACHES,0,1,1,1
+BEACHES,0,2,1,2
+BEACHES,0,3,1,3
+BEACHES,0,4,1,4
+BEACHES,0,5,1,5
+BEACHES,0,6,1,6
+BEACHES,0,7,1,7
+BEACHES,0,8,1,8
+BEACHES,0,9,1,9
+BEACHES,0,10,1,10
+BEACHES,0,11,1,11
+BEACHES,0,12,1,12
+BEACHES,0,13,1,13
+BEACHES,0,14,1,14
+BEACHES,0,15,1,15
+BEACHES,0,16,1,16
+BEACHES,0,17,1,17
+BEACHES,0,18,1,18
+BEACHES,0,19,1,19
+BEACHES,0,20,1,20
+BEACHES,0,21,1,21
+BEACHES,0,22,1,22
+BEACHES,0,23,1,23
+BEACHES,0,24,1,24
+BEACHES,0,25,1,25
+BEACHES,0,26,1,26
+BEACHES,0,27,1,27
+BEACHES,0,28,1,28
+BEACHES,0,29,1,29
+BEACHES,0,30,1,30
+BEACHES,0,31,1,31
+BEACHES,0,32,1,32
+BEACHES,0,33,1,33
+BEACHES,0,34,1,34
+BEACHES,0,35,1,35
+BEACHES,0,36,1,36
+BEACHES,0,37,1,37
+BEACHES,0,38,1,38
+BEACHES,0,39,1,39
+BEACHES,0,40,1,40
+BEACHES,0,41,1,41
+SURVEYS,1,0,1,0
+SURVEYS,1,1,1,1
+SURVEYS,1,2,1,2
+SURVEYS,1,3,1,3
+SURVEYS,1,4,1,4
+SURVEYS,1,5,1,5
+SURVEYS,1,6,1,6
+SURVEYS,1,7,1,7
+SURVEYS,1,8,1,8
+SURVEYS,1,9,1,9
+SURVEYS,1,10,1,10
+SURVEYS,1,11,1,11
+SURVEYS,1,12,1,12
+SURVEYS,1,13,1,13
+SURVEYS,1,14,1,14
+SURVEYS,1,15,1,15
+SURVEYS,1,16,1,16
+SURVEYS,1,17,1,17
+SURVEYS,1,18,1,18
+SURVEYS,1,19,1,19
+SURVEYS,1,20,1,20
+SURVEYS,1,21,1,21
+SURVEYS,1,22,1,22
+SURVEYS,1,23,1,23
+SURVEYS,1,24,1,24
+SURVEYS,1,25,1,25
+SURVEYS,1,26,1,26
+SURVEYS,1,27,1,27
+SURVEYS,1,28,1,28
+SURVEYS,1,29,1,29
+SURVEYS,1,30,1,30
+SURVEYS,1,31,1,31
+SURVEYS,1,32,1,32
+SURVEYS,1,33,1,33
+SURVEYS,1,34,1,34
+SURVEYS,1,35,1,35
+SURVEYS,1,36,1,36
+SURVEYS,1,37,1,37
+SURVEYS,1,38,1,38
+SURVEYS,1,39,1,39
+SURVEYS,1,40,1,40
+SURVEYS,1,41,1,41
+SURVEYS,1,42,1,42
+SURVEYS,1,43,1,43
+SURVEYS,1,44,1,44
+SURVEYS,1,45,1,45
+SURVEYS,1,46,1,46
+SURVEYS,1,47,1,47
+SURVEYS,1,48,1,48
+SURVEYS,1,49,1,49
+SURVEYS,1,50,1,50
+SURVEYS,1,51,1,51
+SURVEYS,1,52,1,52
+SURVEYS,1,53,1,53
+SURVEYS,1,54,1,54
+SURVEYS,1,55,1,55
+SURVEYS,1,56,1,56
+SURVEYS,1,57,1,57
+ANIMALS,2,0,1,0
+ANIMALS,2,1,1,1
+ANIMALS,2,2,1,2
+ANIMALS,2,3,1,3
+ANIMALS,2,4,1,4
+ANIMALS,2,5,1,5
+ANIMALS,2,6,1,6
+LITTER,3,0,1,0
+LITTER,3,1,1,1
+LITTER,3,2,1,2
+LITTER,3,3,1,3
+LITTER,3,4,1,4
+LITTER,3,5,1,5
+LITTER,3,6,1,6
