+++ examples/beachlitter/models/Model_params_another_sheet.csv
@@ -1,0 +1,119 @@
+SHEETS,0,1
+SHEETS,1,2
+SHEETS,2,2
+SHEETS,3,3
+BEACHES,0,0,3,0
+BEACHES,0,1,3,3
+BEACHES,0,2,3,1
+BEACHES,0,3,0,-1
+BEACHES,0,4,3,16
+BEACHES,0,5,0,-1
+BEACHES,0,6,0,-1
+BEACHES,0,7,0,-1
+BEACHES,0,8,0,-1
+BEACHES,0,9,0,-1
+BEACHES,0,10,0,-1
+BEACHES,0,11,3,6
+BEACHES,0,12,3,7
+BEACHES,0,13,3,8
+BEACHES,0,14,3,11
+BEACHES,0,15,3,12
+BEACHES,0,16,3,15
+BEACHES,0,17,3,9
+BEACHES,0,18,0,-1
+BEACHES,0,19,0,-1
+BEACHES,0,20,0,-1
+BEACHES,0,21,0,-1
+BEACHES,0,22,3,17
+BEACHES,0,23,3,18
+BEACHES,0,24,3,19
+BEACHES,0,25,3,20
+BEACHES,0,26,0,-1
+BEACHES,0,27,0,-1
+BEACHES,0,28,3,23
+BEACHES,0,29,3,24
+BEACHES,0,30,0,-1
+BEACHES,0,31,0,-1
+BEACHES,0,32,0,-1
+BEACHES,0,33,0,-1
+BEACHES,0,34,3,29
+BEACHES,0,35,3,57
+BEACHES,0,36,3,58
+BEACHES,0,37,3,56
+BEACHES,0,38,0,-1
+BEACHES,0,39,3,59
+BEACHES,0,40,3,60
+BEACHES,0,41,0,-1
+SURVEYS,1,0,3,0
+SURVEYS,1,1,3,1
+SURVEYS,1,2,0,-1
+SURVEYS,1,3,0,-1
+SURVEYS,1,4,3,2
+SURVEYS,1,5,0,-1
+SURVEYS,1,6,0,-1
+SURVEYS,1,7,0,-1
+SURVEYS,1,8,3,6
+SURVEYS,1,9,3,7
+SURVEYS,1,10,3,8
+SURVEYS,1,11,3,9
+SURVEYS,1,12,0,-1
+SURVEYS,1,13,0,-1
+SURVEYS,1,14,0,-1
+SURVEYS,1,15,0,-1
+SURVEYS,1,16,0,-1
+SURVEYS,1,17,0,-1
+SURVEYS,1,18,0,-1
+SURVEYS,1,19,0,-1
+SURVEYS,1,20,0,-1
+SURVEYS,1,21,0,-1
+SURVEYS,1,22,0,-1
+SURVEYS,1,23,0,-1
+SURVEYS,1,24,0,-1
+SURVEYS,1,25,0,-1
+SURVEYS,1,26,0,-1
+SURVEYS,1,27,0,-1
+SURVEYS,1,28,0,-1
+SURVEYS,1,29,0,-1
+SURVEYS,1,30,0,-1
+SURVEYS,1,31,0,-1
+SURVEYS,1,32,0,-1
+SURVEYS,1,33,0,-1
+SURVEYS,1,34,0,-1
+SURVEYS,1,35,0,-1
+SURVEYS,1,36,0,-1
+SURVEYS,1,37,0,-1
+SURVEYS,1,38,0,-1
+SURVEYS,1,39,0,-1
+SURVEYS,1,40,0,-1
+SURVEYS,1,41,0,-1
+SURVEYS,1,42,0,-1
+SURVEYS,1,43,0,-1
+SURVEYS,1,44,0,-1
+SURVEYS,1,45,0,-1
+SURVEYS,1,46,0,-1
+SURVEYS,1,47,0,-1
+SURVEYS,1,48,0,-1
+SURVEYS,1,49,0,-1
+SURVEYS,1,50,0,-1
+SURVEYS,1,51,0,-1
+SURVEYS,1,52,0,-1
+SURVEYS,1,53,0,-1
+SURVEYS,1,54,0,-1
+SURVEYS,1,55,0,-1
+SURVEYS,1,56,0,-1
+SURVEYS,1,57,0,-1
+ANIMALS,1,0,3,1
+ANIMALS,1,1,3,18
+ANIMALS,1,2,3,16
+ANIMALS,1,3,3,18
+ANIMALS,1,4,0,-1
+ANIMALS,1,5,0,-1
+ANIMALS,1,6,0,-1
+LITTER,2,0,3,0
+LITTER,2,1,0,-1
+LITTER,2,2,3,1
+LITTER,2,3,3,-1
+LITTER,2,4,3,-1
+LITTER,2,5,3,2
+LITTER,2,6,0,-1
+ANOTHER_SHEET,1,3,1,0,1,-1
