+++ examples/beachlitter/models/Model_params_pivoting.csv
@@ -1,0 +1,119 @@
+SHEETS,0,1
+SHEETS,1,1
+SHEETS,2,1
+SHEETS,3,1
+BEACHES,0,0,1,0
+BEACHES,0,1,1,1
+BEACHES,0,2,1,2
+BEACHES,0,3,0,-1
+BEACHES,0,4,1,4
+BEACHES,0,5,0,-1
+BEACHES,0,6,0,-1
+BEACHES,0,7,0,-1
+BEACHES,0,8,0,-1
+BEACHES,0,9,0,-1
+BEACHES,0,10,0,-1
+BEACHES,0,11,0,-1
+BEACHES,0,12,0,-1
+BEACHES,0,13,0,-1
+BEACHES,0,14,1,7
+BEACHES,0,15,1,8
+BEACHES,0,16,0,-1
+BEACHES,0,17,0,-1
+BEACHES,0,18,0,-1
+BEACHES,0,19,0,-1
+BEACHES,0,20,0,-1
+BEACHES,0,21,0,-1
+BEACHES,0,22,0,-1
+BEACHES,0,23,0,-1
+BEACHES,0,24,0,-1
+BEACHES,0,25,0,-1
+BEACHES,0,26,0,-1
+BEACHES,0,27,0,-1
+BEACHES,0,28,0,-1
+BEACHES,0,29,0,-1
+BEACHES,0,30,0,-1
+BEACHES,0,31,0,-1
+BEACHES,0,32,0,-1
+BEACHES,0,33,0,-1
+BEACHES,0,34,0,-1
+BEACHES,0,35,0,-1
+BEACHES,0,36,0,-1
+BEACHES,0,37,0,-1
+BEACHES,0,38,0,-1
+BEACHES,0,39,0,-1
+BEACHES,0,40,0,-1
+BEACHES,0,41,0,-1
+SURVEYS,0,0,1,0
+SURVEYS,0,1,1,3
+SURVEYS,0,2,0,-1
+SURVEYS,0,3,0,-1
+SURVEYS,0,4,1,4
+SURVEYS,0,5,0,-1
+SURVEYS,0,6,0,-1
+SURVEYS,0,7,0,-1
+SURVEYS,0,8,1,7
+SURVEYS,0,9,1,8
+SURVEYS,0,10,1,9
+SURVEYS,0,11,1,10
+SURVEYS,0,12,0,-1
+SURVEYS,0,13,0,-1
+SURVEYS,0,14,0,-1
+SURVEYS,0,15,0,-1
+SURVEYS,0,16,0,-1
+SURVEYS,0,17,0,-1
+SURVEYS,0,18,0,-1
+SURVEYS,0,19,0,-1
+SURVEYS,0,20,0,-1
+SURVEYS,0,21,0,-1
+SURVEYS,0,22,0,-1
+SURVEYS,0,23,0,-1
+SURVEYS,0,24,0,-1
+SURVEYS,0,25,0,-1
+SURVEYS,0,26,0,-1
+SURVEYS,0,27,0,-1
+SURVEYS,0,28,0,-1
+SURVEYS,0,29,0,-1
+SURVEYS,0,30,0,-1
+SURVEYS,0,31,0,-1
+SURVEYS,0,32,0,-1
+SURVEYS,0,33,0,-1
+SURVEYS,0,34,0,-1
+SURVEYS,0,35,0,-1
+SURVEYS,0,36,0,-1
+SURVEYS,0,37,0,-1
+SURVEYS,0,38,0,-1
+SURVEYS,0,39,0,-1
+SURVEYS,0,40,0,-1
+SURVEYS,0,41,0,-1
+SURVEYS,0,42,0,-1
+SURVEYS,0,43,0,-1
+SURVEYS,0,44,0,-1
+SURVEYS,0,45,0,-1
+SURVEYS,0,46,0,-1
+SURVEYS,0,47,0,-1
+SURVEYS,0,48,0,-1
+SURVEYS,0,49,0,-1
+SURVEYS,0,50,0,-1
+SURVEYS,0,51,0,-1
+SURVEYS,0,52,0,-1
+SURVEYS,0,53,0,-1
+SURVEYS,0,54,0,-1
+SURVEYS,0,55,0,-1
+SURVEYS,0,56,0,-1
+SURVEYS,0,57,0,-1
+ANIMALS,0,0,0,-1
+ANIMALS,0,1,0,-1
+ANIMALS,0,2,0,-1
+ANIMALS,0,3,0,-1
+ANIMALS,0,4,0,-1
+ANIMALS,0,5,0,-1
+ANIMALS,0,6,0,-1
+LITTER,0,0,1,3
+LITTER,0,1,1,6
+LITTER,0,2,0,11
+LITTER,0,3,0,-1
+LITTER,0,4,0,-1
+LITTER,0,5,1,11
+LITTER,0,6,0,-1
+ANOTHER_SHEET,1,1,1,0,1,2
